Moved models into session blaze module and started rewriting user session ext
