+++ src/utils/mod.rs
@@ -1,6 +1,5 @@
 pub mod components;
 pub mod hashing;
 pub mod logging;
-pub mod models;
 pub mod parsing;
 pub mod types;
